+++ src/gui/plugins/modules/EntityContextMenu.cc
@@ -202,7 +202,7 @@ void EntityContextMenu::OnRequest(const QString &_request, const QString &_data)
   }
   else if (request == "view_transparent")
   {
-    ignition::msgs::StringMsg req;
+    msgs::StringMsg req;
     req.set_data(_data.toStdString());
     this->dataPtr->node.Request(this->dataPtr->viewTransparentService, req, cb);
   }