+++ src/gui/plugins/view_angle/ViewAngle.cc
@@ -61,18 +61,33 @@ namespace ignition::gazebo
     /// \brief View Control service name
     public: std::string viewControlService;
 
+    /// \brief View Control reference visual service name
+    public: std::string viewControlRefVisualService;
+
+    /// \brief View Control sensitivity service name
+    public: std::string viewControlSensitivityService;
+
     /// \brief Move gui camera to pose service name
     public: std::string moveToPoseService;
 
-    /// \brief gui camera pose
-    public: math::Pose3d camPose;
+    /// \brief Move gui camera to model service name
+    public: std::string moveToModelService;
+
+    /// \brief New move to model message
+    public: bool newMoveToModel = false;
+
+    /// \brief Distance of the camera to the model
+    public: double distanceMoveToModel = 0.0;
 
     /// \brief Camera horizontal fov
-    public: double horizontal_fov = 0.0;
+    public: double horizontalFov = 0.0;
 
     /// \brief Flag indicating if there is a new camera horizontalFOV
     /// coming from qml side
     public: bool newHorizontalFOV = false;
+
+    /// \brief gui camera pose
+    public: math::Pose3d camPose;
 
     /// \brief GUI camera near/far clipping distance (QList order is near, far)
     public: QList<double> camClipDist{0.0, 0.0};
@@ -85,6 +100,14 @@ namespace ignition::gazebo
     /// used to update qml side
     /// \return True if there is a new clipping distance from gui camera
     public: bool UpdateQtCamClipDist();
+
+    /// \brief View Control type
+    public: rendering::CameraProjectionType viewControlType =
+              rendering::CameraProjectionType::CPT_PERSPECTIVE;
+
+    /// \brief Checks if there is a new view controller, used to update qml side
+    /// \return True if there is a new view controller from gui camera
+    public: bool UpdateQtViewControl();
 
     /// \brief Checks if there is new camera horizontal fov from gui camera,
     /// used to update qml side
@@ -153,6 +176,14 @@ void ViewAngle::LoadConfig(const tinyxml2::XMLElement *_pluginElem)
   // view control requests
   this->dataPtr->viewControlService = "/gui/camera/view_control";
 
+  // view control reference visual requests
+  this->dataPtr->viewControlRefVisualService =
+      "/gui/camera/view_control/reference_visual";
+
+  // view control sensitivity requests
+  this->dataPtr->viewControlSensitivityService =
+      "/gui/camera/view_control/sensitivity";
+
   // Subscribe to camera pose
   std::string topic = "/gui/camera/pose";
   this->dataPtr->node.Subscribe(
@@ -160,6 +191,13 @@ void ViewAngle::LoadConfig(const tinyxml2::XMLElement *_pluginElem)
 
   // Move to pose service
   this->dataPtr->moveToPoseService = "/gui/move_to/pose";
+
+  // Move to model service
+  this->dataPtr->moveToModelService = "/gui/move_to/model";
+  this->dataPtr->node.Advertise(this->dataPtr->moveToModelService,
+      &ViewAngle::OnMoveToModelService, this);
+  ignmsg << "Move to model service on ["
+         << this->dataPtr->moveToModelService << "]" << std::endl;
 
   ignition::gui::App()->findChild<
     ignition::gui::MainWindow *>()->installEventFilter(this);
@@ -177,10 +215,16 @@ bool ViewAngle::eventFilter(QObject *_obj, QEvent *_event)
     {
       this->CamClipDistChanged();
     }
+
     // updates qml cam horizontal fov spin boxes if changed
     if (this->dataPtr->UpdateQtCamHorizontalFOV())
     {
       this->CamHorizontalFOVChanged();
+    }
+
+    if (this->dataPtr->UpdateQtViewControl())
+    {
+      this->ViewControlIndexChanged();
     }
   }
   else if (_event->type() ==
@@ -266,6 +310,46 @@ void ViewAngle::OnViewControl(const QString &_controller)
 }
 
 /////////////////////////////////////////////////
+void ViewAngle::OnViewControlReferenceVisual(bool _enable)
+{
+  std::function<void(const msgs::Boolean &, const bool)> cb =
+      [](const msgs::Boolean &/*_rep*/, const bool _result)
+  {
+    if (!_result)
+      ignerr << "Error setting view controller reference visual" << std::endl;
+  };
+
+  msgs::Boolean req;
+  req.set_data(_enable);
+
+  this->dataPtr->node.Request(
+      this->dataPtr->viewControlRefVisualService, req, cb);
+}
+
+/////////////////////////////////////////////////
+void ViewAngle::OnViewControlSensitivity(double _sensitivity)
+{
+  std::function<void(const msgs::Boolean &, const bool)> cb =
+      [](const msgs::Boolean &/*_rep*/, const bool _result)
+  {
+    if (!_result)
+      ignerr << "Error setting view controller sensitivity" << std::endl;
+  };
+
+  if (_sensitivity <= 0.0)
+  {
+    ignerr << "View controller sensitivity must be greater than 0" << std::endl;
+    return;
+  }
+
+  msgs::Double req;
+  req.set_data(_sensitivity);
+
+  this->dataPtr->node.Request(
+      this->dataPtr->viewControlSensitivityService, req, cb);
+}
+
+/////////////////////////////////////////////////
 QList<double> ViewAngle::CamPose() const
 {
   return QList({
@@ -307,6 +391,84 @@ void ViewAngle::SetCamPose(double _x, double _y, double _z,
 }
 
 /////////////////////////////////////////////////
+bool ViewAngle::OnMoveToModelService(const ignition::msgs::GUICamera &_msg,
+  ignition::msgs::Boolean &_res)
+{
+  std::lock_guard<std::mutex> lock(this->dataPtr->mutex);
+  auto scene = this->dataPtr->camera->Scene();
+
+  auto visualToMove = scene->VisualByName(_msg.name());
+  if (nullptr == visualToMove)
+  {
+    ignerr << "Failed to get visual with ID ["
+           << _msg.name() << "]" << std::endl;
+    _res.set_data(false);
+    return false;
+  }
+  Entity entityId = kNullEntity;
+  try
+  {
+    // TODO(ahcorde): When forward porting this to Garder change var type to
+    // unsigned int
+    entityId = std::get<int>(visualToMove->UserData("gazebo-entity"));
+  }
+  catch(std::bad_variant_access &_e)
+  {
+    ignerr << "Failed to get gazebo-entity user data ["
+           << visualToMove->Name() << "]" << std::endl;
+    _res.set_data(false);
+    return false;
+  }
+
+  math::Quaterniond q(
+    _msg.pose().orientation().w(),
+    _msg.pose().orientation().x(),
+    _msg.pose().orientation().y(),
+    _msg.pose().orientation().z());
+
+  ignition::math::Vector3d axis;
+  double angle;
+  q.ToAxis(axis, angle);
+
+  std::function<void(const msgs::Boolean &, const bool)> cb =
+      [](const msgs::Boolean &/*_rep*/, const bool _result)
+  {
+    if (!_result)
+      ignerr << "Error setting view controller" << std::endl;
+  };
+
+  msgs::StringMsg req;
+  std::string str = _msg.projection_type();
+  if (str.find("Orbit") != std::string::npos ||
+      str.find("orbit") != std::string::npos)
+  {
+    req.set_data("orbit");
+  }
+  else if (str.find("Ortho") != std::string::npos ||
+           str.find("ortho") != std::string::npos )
+  {
+    req.set_data("ortho");
+  }
+  else
+  {
+    ignerr << "Unknown view controller selected: " << str << std::endl;
+    _res.set_data(false);
+    return false;
+  }
+
+  this->dataPtr->node.Request(this->dataPtr->viewControlService, req, cb);
+
+  this->dataPtr->viewingAngle = true;
+  this->dataPtr->newMoveToModel = true;
+  this->dataPtr->viewAngleDirection = axis;
+  this->dataPtr->distanceMoveToModel = _msg.pose().position().z();
+  this->dataPtr->selectedEntities.push_back(entityId);
+
+  _res.set_data(true);
+  return true;
+}
+
+/////////////////////////////////////////////////
 void ViewAngle::CamPoseCb(const msgs::Pose &_msg)
 {
   std::lock_guard<std::mutex> lock(this->dataPtr->mutex);
@@ -322,13 +484,13 @@ void ViewAngle::CamPoseCb(const msgs::Pose &_msg)
 /////////////////////////////////////////////////
 double ViewAngle::HorizontalFOV() const
 {
-  return this->dataPtr->horizontal_fov;
+  return this->dataPtr->horizontalFov;
 }
 
 /////////////////////////////////////////////////
 void ViewAngle::SetHorizontalFOV(double _horizontalFOV)
 {
-  this->dataPtr->horizontal_fov = _horizontalFOV;
+  this->dataPtr->horizontalFov = _horizontalFOV;
   this->dataPtr->newHorizontalFOV = true;
 }
 
@@ -470,7 +632,7 @@ void ViewAnglePrivate::OnRender()
   // Camera horizontalFOV
   if (this->newHorizontalFOV)
   {
-    this->camera->SetHFOV(gz::math::Angle(this->horizontal_fov));
+    this->camera->SetHFOV(gz::math::Angle(this->horizontalFov));
     this->newHorizontalFOV = false;
   }
 }
@@ -480,15 +642,44 @@ void ViewAnglePrivate::OnComplete()
 {
   this->viewingAngle = false;
   this->moveToPoseValue.reset();
+  if (this->newMoveToModel)
+  {
+    this->selectedEntities.pop_back();
+    this->newMoveToModel = false;
+
+    auto cameraPose = this->camera->WorldPose();
+    auto distance = -(this->viewAngleDirection * this->distanceMoveToModel);
+
+    if (!math::equal(this->viewAngleDirection.X(), 0.0))
+    {
+      cameraPose.Pos().X(distance.X());
+    }
+    if (!math::equal(this->viewAngleDirection.Y(), 0.0))
+    {
+      cameraPose.Pos().Y(distance.Y());
+    }
+    if (!math::equal(this->viewAngleDirection.Z(), 0.0))
+    {
+      cameraPose.Pos().Z(distance.Z());
+    }
+
+    this->moveToPoseValue = {
+      cameraPose.Pos().X(),
+      cameraPose.Pos().Y(),
+      cameraPose.Pos().Z(),
+      cameraPose.Rot().Roll(),
+      cameraPose.Rot().Pitch(),
+      cameraPose.Rot().Yaw()};
+  }
 }
 
 /////////////////////////////////////////////////
 bool ViewAnglePrivate::UpdateQtCamHorizontalFOV()
 {
   bool updated = false;
-  if (std::abs(this->camera->HFOV().Radian() - this->horizontal_fov) > 0.0001)
+  if (std::abs(this->camera->HFOV().Radian() - this->horizontalFov) > 0.0001)
   {
-    this->horizontal_fov = this->camera->HFOV().Radian();
+    this->horizontalFov = this->camera->HFOV().Radian();
     updated = true;
   }
   return updated;
@@ -510,6 +701,31 @@ bool ViewAnglePrivate::UpdateQtCamClipDist()
     updated = true;
   }
   return updated;
+}
+
+/////////////////////////////////////////////////
+int ViewAngle::ViewControlIndex() const
+{
+  if (this->dataPtr->viewControlType ==
+        rendering::CameraProjectionType::CPT_PERSPECTIVE)
+    return 0;
+
+  return 1;
+}
+
+/////////////////////////////////////////////////
+bool ViewAnglePrivate::UpdateQtViewControl()
+{
+  if (!this->camera)
+    return false;
+
+  if (this->camera->ProjectionType() != this->viewControlType)
+  {
+    this->viewControlType = this->camera->ProjectionType();
+    return true;
+  }
+
+  return false;
 }
 
 // Register this plugin