+++ src/gui/plugins/view_angle/ViewAngle.hh
@@ -19,6 +19,8 @@
 #define IGNITION_GAZEBO_GUI_VIEWANGLE_HH_
 
 #include <ignition/msgs/pose.pb.h>
+#include <ignition/msgs/boolean.pb.h>
+#include <ignition/msgs/gui_camera.pb.h>
 
 #include <memory>
 
@@ -54,6 +56,13 @@ namespace gazebo
       NOTIFY CamClipDistChanged
     )
 
+    /// \brief view controller index for qml side (0: orbit; 1: ortho)
+    Q_PROPERTY(
+      int viewControlIndex
+      READ ViewControlIndex
+      NOTIFY ViewControlIndexChanged
+    )
+
     /// \brief gui camera horizontal fov
     Q_PROPERTY(
       double horizontalFOV
@@ -86,8 +95,19 @@ namespace gazebo
     /// \param[in] _mode New camera view controller
     public slots: void OnViewControl(const QString &_controller);
 
-    /// \brief Get the current gui camera pose.
-    public: Q_INVOKABLE QList<double> CamPose() const;
+    /// \brief Callback in Qt thread when camera view reference visual state
+    /// changes.
+    /// \param[in] _enable True to enable camera view control reference visual,
+    /// false to hide it
+    public slots: void OnViewControlReferenceVisual(bool _enable);
+
+    /// \brief Callback in Qt thread when camera view controller changes.
+    /// \param[in] _sensitivity View control sensitivity vlaue
+    public slots: void OnViewControlSensitivity(double _sensitivity);
+
+    /// \brief Updates gui camera's Horizontal fov
+    /// \param[in] _horizontalFOV Horizontal fov
+    public slots: void SetHorizontalFOV(double _horizontalFOV);
 
     /// \brief Get the current gui horizontal fov.
     public: Q_INVOKABLE double HorizontalFOV() const;
@@ -95,12 +115,11 @@ namespace gazebo
     /// \brief Notify that the gui camera's horizontal fov changed
     signals: void CamHorizontalFOVChanged();
 
+    /// \brief Get the current gui camera pose.
+    public: Q_INVOKABLE QList<double> CamPose() const;
+
     /// \brief Notify that the gui camera pose has changed.
     signals: void CamPoseChanged();
-
-    /// \brief Updates gui camera's Horizontal fov
-    /// \param[in] _horizontalFOV Horizontal fov
-    public slots: void SetHorizontalFOV(double _horizontalFOV);
 
     /// \brief Callback to update gui camera pose
     /// \param[in] _x, _y, _z cartesion coordinates
@@ -112,6 +131,12 @@ namespace gazebo
     /// \param[in] _msg Pose message
     public: void CamPoseCb(const msgs::Pose &_msg);
 
+    /// \brief Move to model service received
+    /// \param[in] _msg GUI camera message
+    /// \param[in] _res Response
+    public: bool OnMoveToModelService(const ignition::msgs::GUICamera &_msg,
+      ignition::msgs::Boolean &_res);
+
     /// \brief Get the current gui camera's near and far clipping distances
     public: Q_INVOKABLE QList<double> CamClipDist() const;
 
@@ -122,6 +147,12 @@ namespace gazebo
     /// \param[in] _near Near clipping plane distance
     /// \param[in] _far Far clipping plane distance
     public slots: void SetCamClipDist(double _near, double _far);
+
+    /// \brief Get the current index for view controller (on qml side)
+    public: int ViewControlIndex() const;
+
+    /// \brief Notify that the camera's view controller has changed
+    signals: void ViewControlIndexChanged();
 
     /// \internal
     /// \brief Pointer to private data.