+++ src/systems/thruster/Thruster.cc
@@ -29,10 +29,12 @@
 
 #include "ignition/gazebo/components/AngularVelocity.hh"
 #include "ignition/gazebo/components/BatterySoC.hh"
+#include "ignition/gazebo/components/BatteryPowerLoad.hh"
 #include "ignition/gazebo/components/ChildLinkName.hh"
 #include "ignition/gazebo/components/JointAxis.hh"
 #include "ignition/gazebo/components/JointVelocityCmd.hh"
 #include "ignition/gazebo/components/LinearVelocity.hh"
+#include "ignition/gazebo/components/Name.hh"
 #include "ignition/gazebo/components/Pose.hh"
 #include "ignition/gazebo/components/World.hh"
 #include "ignition/gazebo/Link.hh"
@@ -64,6 +66,9 @@ class ignition::gazebo::systems::ThrusterPrivateData
 
   /// \brief The link entity which will spin
   public: ignition::gazebo::Entity linkEntity;
+
+  /// \brief Battery consumer entity
+  public: Entity consumerEntity;
 
   /// \brief Axis along which the propeller spins. Expressed in the joint
   /// frame. Addume this doesn't change during simulation.
@@ -125,6 +130,15 @@ class ignition::gazebo::systems::ThrusterPrivateData
 
   /// \brief Topic name used to control thrust.
   public: std::string topic = "";
+
+  /// \brief Battery entity used by the thruster to consume power.
+  public: std::string batteryName = "";
+
+  /// \brief Battery power load of the thruster.
+  public: double powerLoad = 0.0;
+
+  /// \brief Has the battery consumption being initialized.
+  public: bool batteryInitialized = false;
 
   /// \brief Callback for handling thrust update
   public: void OnCmdThrust(const msgs::Double &_msg);
@@ -363,6 +377,22 @@ void Thruster::Configure(
   {
     igndbg << "Using velocity control for propeller joint." << std::endl;
   }
+
+  // Get power load and battery name info
+  if (_sdf->HasElement("power_load"))
+  {
+    if (!_sdf->HasElement("battery_name"))
+    {
+      ignerr << "Specified a <power_load> but missing <battery_name>."
+          "Specify a battery name so the power load can be assigned to it."
+          << std::endl;
+    }
+    else
+    {
+      this->dataPtr->powerLoad = _sdf->Get<double>("power_load");
+      this->dataPtr->batteryName = _sdf->Get<std::string>("battery_name");
+    }
+  }
 }
 
 /////////////////////////////////////////////////
@@ -441,6 +471,52 @@ void Thruster::PreUpdate(
   if (!this->dataPtr->enabled)
   {
     return;
+  }
+
+  // Intit battery consumption if it was set
+  if (!this->dataPtr->batteryName.empty() &&
+      !this->dataPtr->batteryInitialized)
+  {
+    this->dataPtr->batteryInitialized = true;
+
+    // Check that a battery exists with the specified name
+    Entity batteryEntity;
+    int numBatteriesWithName = 0;
+    _ecm.Each<components::BatterySoC, components::Name>(
+      [&](const Entity &_entity,
+        const components::BatterySoC */*_BatterySoC*/,
+        const components::Name *_name)->bool
+      {
+        if (this->dataPtr->batteryName == _name->Data())
+        {
+          ++numBatteriesWithName;
+          batteryEntity = _entity;
+        }
+        return true;
+      });
+    if (numBatteriesWithName == 0)
+    {
+      ignerr << "Can't assign battery consumption to battery: ["
+             << this->dataPtr->batteryName << "]. No batteries"
+             "were found with the given name." << std::endl;
+      return;
+    }
+    if (numBatteriesWithName > 1)
+    {
+      ignerr << "More than one battery found with name: ["
+             << this->dataPtr->batteryName << "]. Please make"
+             "sure battery names are unique within the system."
+             << std::endl;
+      return;
+    }
+
+    // Create the battery consumer entity and its component
+    this->dataPtr->consumerEntity = _ecm.CreateEntity();
+    components::BatteryPowerLoadInfo batteryPowerLoadInfo{
+        batteryEntity, this->dataPtr->powerLoad};
+    _ecm.CreateComponent(this->dataPtr->consumerEntity,
+        components::BatteryPowerLoad(batteryPowerLoadInfo));
+    _ecm.SetParentEntity(this->dataPtr->consumerEntity, batteryEntity);
   }
 
   ignition::gazebo::Link link(this->dataPtr->linkEntity);