+++ test/integration/scene_broadcaster_system.cc
@@ -64,7 +64,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(PoseInfo))
   gazebo::Server server(serverConfig);
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
-  EXPECT_EQ(24u, *server.EntityCount());
+  EXPECT_EQ(25u, *server.EntityCount());
 
   // Create pose subscriber
   transport::Node node;
@@ -76,7 +76,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(PoseInfo))
     ASSERT_TRUE(_msg.header().has_stamp());
     EXPECT_LT(0, _msg.header().stamp().sec() +  _msg.header().stamp().nsec());
 
-    EXPECT_EQ(16, _msg.pose_size());
+    EXPECT_EQ(17, _msg.pose_size());
 
     std::map<int, std::string> entityMap;
     for (auto p = 0; p < _msg.pose_size(); ++p)
@@ -84,7 +84,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(PoseInfo))
       entityMap.insert(std::make_pair(_msg.pose(p).id(), _msg.pose(p).name()));
     }
 
-    EXPECT_EQ(16u, entityMap.size());
+    EXPECT_EQ(17u, entityMap.size());
 
     received = true;
   };
@@ -114,7 +114,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(SceneInfo))
   gazebo::Server server(serverConfig);
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
-  EXPECT_EQ(24u, *server.EntityCount());
+  EXPECT_EQ(25u, *server.EntityCount());
 
   // Run server
   server.Run(true, 1, false);
@@ -160,7 +160,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(SceneGraph))
   gazebo::Server server(serverConfig);
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
-  EXPECT_EQ(24u, *server.EntityCount());
+  EXPECT_EQ(25u, *server.EntityCount());
 
   // Run server
   server.Run(true, 1, false);
@@ -200,7 +200,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(SceneTopic))
   gazebo::Server server(serverConfig);
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
-  EXPECT_EQ(24u, *server.EntityCount());
+  EXPECT_EQ(25u, *server.EntityCount());
 
   // Create requester
   transport::Node node;
@@ -297,7 +297,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(DeletedTopic))
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
 
-  const std::size_t initEntityCount = 24;
+  const std::size_t initEntityCount = 25;
   EXPECT_EQ(initEntityCount, *server.EntityCount());
 
   // Subscribe to deletions
@@ -358,7 +358,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(SpawnedModel))
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
 
-  const std::size_t initEntityCount = 24;
+  const std::size_t initEntityCount = 25;
   EXPECT_EQ(initEntityCount, *server.EntityCount());
 
   server.Run(true, 1, false);
@@ -427,7 +427,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(State))
   gazebo::Server server(serverConfig);
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
-  EXPECT_EQ(24u, *server.EntityCount());
+  EXPECT_EQ(25u, *server.EntityCount());
   transport::Node node;
 
   // Run server
@@ -457,7 +457,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(State))
       [&](const msgs::SerializedStepMap &_res, const bool _success)
   {
     EXPECT_TRUE(_success);
-    checkMsg(_res, 24);
+    checkMsg(_res, 25);
   };
   std::function<void(const msgs::SerializedStepMap &)> cb2 =
       [&](const msgs::SerializedStepMap &_res)
@@ -469,7 +469,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(State))
   std::function<void(const msgs::SerializedStepMap &)> cbAsync =
       [&](const msgs::SerializedStepMap &_res)
   {
-    checkMsg(_res, 24);
+    checkMsg(_res, 25);
   };
 
   // The request is blocking even though it's meant to be async, so we spin a
@@ -538,7 +538,7 @@ TEST_P(SceneBroadcasterTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(StateStatic))
   gazebo::Server server(serverConfig);
   EXPECT_FALSE(server.Running());
   EXPECT_FALSE(*server.Running(0));
-  EXPECT_EQ(8u, *server.EntityCount());
+  EXPECT_EQ(9u, *server.EntityCount());
   transport::Node node;
 
   // Run server
@@ -656,6 +656,7 @@ TEST_P(SceneBroadcasterTest,
       // remove a component from an entity
       if (_info.iterations == 2)
       {
+        std::vector<gazebo::Entity> entitiesToRemoveFrom;
         _ecm.Each<ignition::gazebo::components::Model,
                   ignition::gazebo::components::Name,
                   ignition::gazebo::components::Pose>(
@@ -666,10 +667,14 @@ TEST_P(SceneBroadcasterTest,
           {
             if (_name->Data() == "box")
             {
-              _ecm.RemoveComponent<ignition::gazebo::components::Pose>(_entity);
+              entitiesToRemoveFrom.push_back(_entity);
             }
             return true;
           });
+        for (const auto &entity : entitiesToRemoveFrom)
+        {
+          _ecm.RemoveComponent<gazebo::components::Pose>(entity);
+        }
       }
       // add a component to an entity
       else if (_info.iterations == 3)
@@ -952,10 +957,12 @@ TEST_P(SceneBroadcasterTest,
   EXPECT_TRUE(result);
 
   ASSERT_TRUE(res.has_ambient());
-  EXPECT_EQ(math::Color(1.0, 1.0, 1.0, 1.0), msgs::Convert(res.ambient()));
+  EXPECT_EQ(math::Color(1.0f, 1.0f, 1.0f, 1.0f),
+      msgs::Convert(res.ambient()));
 
   ASSERT_TRUE(res.has_background());
-  EXPECT_EQ(math::Color(0.8, 0.8, 0.8, 1.0), msgs::Convert(res.background()));
+  EXPECT_EQ(math::Color(0.8f, 0.8f, 0.8f, 1.0f),
+      msgs::Convert(res.background()));
 
   EXPECT_TRUE(res.shadows());
   EXPECT_FALSE(res.grid());